+++ tests/Pipaslot.Mediator.Tests/E2E/ErrorHandling/NoHandlerTests.cs
@@ -2,7 +2,7 @@
 using Pipaslot.Mediator.Tests.InvalidActions;
 using System.Threading.Tasks;
 
-namespace Pipaslot.Mediator.Tests.E2E;
+namespace Pipaslot.Mediator.Tests.E2E.ErrorHandling;
 
 public class NoHandlerTests
 {
