+++ tests/Pipaslot.Mediator.Tests/E2E/Nesting/NestedUnhandledCollapseTests.cs
@@ -7,7 +7,7 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace Pipaslot.Mediator.Tests;
+namespace Pipaslot.Mediator.Tests.E2E.Nesting;
 
 /// <summary>
 /// A handler that makes a nested <see cref="IMediator.DispatchUnhandled"/>/<see cref="IMediator.ExecuteUnhandled"/>
@@ -15,20 +15,20 @@ namespace Pipaslot.Mediator.Tests;
 /// <c>AddError</c>) instead of throwing, or throws a business exception directly. Because the call is nested, a
 /// recorded notification propagates up to the parent context via <see cref="NotificationPropagationMiddleware"/>
 /// before <c>*Unhandled</c> throws for the same failure.
-/// Introduced in version 9.0.0: the parent's own unguarded <c>Dispatch</c>/<c>Execute</c> boundary no longer re-adds a
-/// duplicate generic wrapper message for a <see cref="MediatorUnhandledErrorException"/>, and resolves a typed
+/// The parent's own unguarded <c>Dispatch</c>/<c>Execute</c> boundary must not re-add a duplicate generic wrapper
+/// message for a <see cref="MediatorUnhandledErrorException"/>, and must resolve a typed
 /// <see cref="IMediatorExceptionHandler{TException}"/> for the original exception type when the inner failure
 /// reaches it as a real exception rather than a swallowed one.
 /// Kept in its own file (like  <c>Notifications/NotificationPropagationTests.cs</c>) since these scenarios span
 /// two nested contexts and have no natural home in a single-context E2E test.
 /// </summary>
-public class Mediator_NestedUnhandledCollapseTests
+public class NestedUnhandledCollapseTests
 {
     /// <summary>
     /// The outer handler does not catch the exception from its nested call (mirrors a real-world gap), so it
-    /// bubbles out to the outer <c>Dispatch</c>. Before Unit 4, the outer catch additionally added
+    /// bubbles out to the outer <c>Dispatch</c>. The outer catch must not additionally add
     /// <c>MediatorUnhandledErrorException</c>'s wrapper message as a second, duplicate notification on top of the
-    /// one already propagated by <see cref="NotificationPropagationMiddleware"/> - Unit 4 removes that duplication.
+    /// one already propagated by <see cref="NotificationPropagationMiddleware"/>.
     /// </summary>
     [Fact]
     public async Task Dispatch_NestedDispatchUnhandledFailsWithoutThrow_ParentGetsExactlyThePropagatedNotification()
