+++ tests/Pipaslot.Mediator.Tests/E2E/Outcomes/BasicPassingTests.cs
@@ -1,7 +1,7 @@
 ﻿using Pipaslot.Mediator.Tests.ValidActions;
 using System.Threading.Tasks;
 
-namespace Pipaslot.Mediator.Tests.E2E;
+namespace Pipaslot.Mediator.Tests.E2E.Outcomes;
 
 public class BasicPassingTests
 {
