+++ tests/Pipaslot.Mediator.Tests/Middlewares/Features/MiddlewareParametersFeatureTests.cs
@@ -5,7 +5,7 @@ using System;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace Pipaslot.Mediator.Tests.E2E;
+namespace Pipaslot.Mediator.Tests.Middlewares.Features;
 
 public class MiddlewareParametersFeatureTests
 {
