+++ tests/Pipaslot.Mediator.Tests/Services/HandlerExistenceChecker_MatchActionHandlersTests.cs
@@ -6,7 +6,7 @@ using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace Pipaslot.Mediator.Tests;
+namespace Pipaslot.Mediator.Tests.Services;
 
 public class HandlerExistenceChecker_MatchActionHandlersTests
 {
